Upgrade desktops to stretch

+++ modules/ocf/manifests/puppet.pp
@@ -38,35 +38,57 @@ class ocf::puppet($stage = 'first') {
       minute  => [fqdn_rand(30), fqdn_rand(30) + 30],
     }
   } else {
-    package { ['facter', 'puppet']: }
+    package { ['facter', 'puppet', 'augeas-tools', 'ruby-augeas']: }
 
-    # configure puppet agent
-    # set environment to match server and disable cached catalog on failure
-    augeas { '/etc/puppet/puppet.conf':
-      context => '/files/etc/puppet/puppet.conf',
-      changes => [
-        # These changes can change the puppetmaster config, which is
-        # defined separately in the ocf_puppet module, causing the
-        # puppet agent on the puppetmaster to restart twice. Make sure
-        # the changes made here are also made in that module.
-        "set agent/environment ${::environment}",
-        'set agent/usecacheonfailure false',
-        'set main/pluginsync true',
-        'set main/stringify_facts false',
-        'set main/rundir /run/puppet',
+    if $::lsbdistcodename == 'jessie' {
+      # configure puppet agent
+      # set environment to match server and disable cached catalog on failure
+      augeas { '/etc/puppet/puppet.conf':
+        context => '/files/etc/puppet/puppet.conf',
+        changes => [
+          # These changes can change the puppetmaster config, which is
+          # defined separately in the ocf_puppet module, causing the
+          # puppet agent on the puppetmaster to restart twice. Make sure
+          # the changes made here are also made in that module.
+          "set agent/environment ${::environment}",
+          'set agent/usecacheonfailure false',
+          'set main/pluginsync true',
+          'set main/stringify_facts false',
+          'set main/rundir /run/puppet',
 
-        # future parser breaks too many 3rd-party modules
-        'rm main/parser',
+          # future parser breaks too many 3rd-party modules
+          'rm main/parser',
 
-        # templatedir is deprecated in 3.8+ and we don't use it
-        'rm main/templatedir',
-      ],
-      require => Package['augeas-tools', 'libaugeas-ruby', 'puppet'],
-      notify  => Service['puppet'],
-    }
+          # templatedir is deprecated in 3.8+ and we don't use it
+          'rm main/templatedir',
+        ],
+        require => Package['augeas-tools', 'libaugeas-ruby', 'puppet'],
+        notify  => Service['puppet'],
+      }
 
-    service { 'puppet':
-      require   => Package['puppet'],
+      service { 'puppet':
+        require   => Package['puppet'],
+      }
+    } else {
+      augeas { '/etc/puppet/puppet.conf':
+        context => '/files/etc/puppet/puppet.conf',
+        changes => [
+          "set agent/environment ${::environment}",
+          'set agent/usecacheonfailure false',
+
+          # TODO: Remove this after the puppetserver upgrade is complete
+          'set main/server dev-puppet',
+        ],
+        require => Package['augeas-tools', 'puppet'],
+      }
+
+      # Run puppet as a cron job
+      cron { 'puppet-agent':
+        ensure  => present,
+        command => '/usr/bin/puppet agent --verbose --onetime --no-daemonize --logdest syslog > /dev/null 2>&1',
+        user    => 'root',
+        minute  => [fqdn_rand(30), fqdn_rand(30) + 30],
+      }
     }
   }
 
@@ -82,9 +104,6 @@ class ocf::puppet($stage = 'first') {
       force   => true,
       backup  => false;
   }
-
-  # install augeas
-  package { [ 'augeas-tools', 'libaugeas-ruby']:; }
 
   # install custom scripts
   file {
